% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{add_retro}
\alias{add_retro}
\title{Join nwmRetro data to NHD object}
\usage{
add_retro(obj = NULL, comid = NULL)
}
\arguments{
\item{obj}{a NHD object with a COMID attribute}

\item{comid}{a NHD COMID}
}
\value{
a simple features (sf) object
}
\description{
Joins the retrospective averages to an input NHD object and computes the seasoal, annual and monthly averages.
}
