% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find}
\alias{find}
\title{Find Spatial NHD or WBD data}
\usage{
find(AOI, type = "flowlines", spatial = TRUE)
}
\arguments{
\item{AOI}{a Spatial Object}

\item{type}{the WBD or nhdplus object to return. Options include:
\itemize{
\item HUC08
\item HUC12
\item flowlines
\item catchments
\item waterbodies
}}

\item{spatial}{return `sp` (default) if \code{FALSE} return `sf`}
}
\value{
a \code{Spatial} object
}
\description{
returns a spatail sub set of the NHD of WBD for an input spatial object using the CIDA Web Coverage Services
}
\author{
Mike Johnson
}
