% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwm_averages.R
\name{nwmAverages}
\alias{nwmAverages}
\title{Long term NWM averages}
\usage{
nwmAverages(include.averages = TRUE, include.timezones = TRUE,
  include.months = month.abb)
}
\arguments{
\item{include.averages}{A logical scalar. Should seasonal (fall, spring, summer, winter)
and annual averages be included?}

\item{include.timezones}{A logical scalar. Should timezone names (e.x. "America/New_York")
and hour offsets from UTC be included?}

\item{include.months}{A character vector.  Months to include (e.x.  \code{c("Jan", "Feb")}).}
}
\value{
a \code{data.frame} object
}
\description{
Returns A dataset containing a 23-year run of of the National Water Model
version 1.0 provided by NCAR and OWP
}
\author{
Pat Johnson
}
