% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return.what}
\alias{return.what}
\title{Choose what to return for HydroData Calls}
\usage{
return.what(report)
}
\arguments{
\item{report}{character sting to append values to}
}
\value{
a list of HydroData components
}
\description{
A function defining what should be returned in a HydroData object
}
