% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwmRetro_data.R
\docType{data}
\name{timezones}
\alias{timezones}
\title{Timezones}
\format{a data.frame with timezones and offsets}
\usage{
timezones
}
\description{
Dataset of timezones
}
\examples{
\dontrun{
nwmRetro::nwm_averages
}
}
\keyword{datasets}
