% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write Spatial Object}
\usage{
write(obj, file = stop("'file' must be specified"), layer = NULL)
}
\arguments{
\item{obj}{a spatial or simple features object}

\item{file}{a path to the file name to be created. Must include an extension. Valid options include:
\itemize{
  \item shp
  \item csv
  \item gpkg
  \item geojson
  \item sqlite
  \item xlsx
  \item rda
}}

\item{layer}{layer name (varies by driver); if layer is missing, the basename of file is taken without the extension.}
}
\value{
a file saved to the path given
}
\description{
Save a spatial file to disk in a varitey of formats, a simplfied wrapper around \code{st_write}
}
\author{
Mike Johnson
}
