% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getExtension}
\alias{getExtension}
\title{Parse Files Extension}
\usage{
getExtension(path)
}
\arguments{
\item{path}{a path to a file}
}
\value{
a character string of the associated file extension
}
\description{
Get a file extension from Path
}
