% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join nwmRetro data to NHD object}
\usage{
<<<<<<< HEAD
join(obj)
=======
join(obj = NULL)
>>>>>>> cabb7542fda948e253be880263e733cc398e4e85
}
\arguments{
\item{obj}{a NHD object with a COMID attribute}
}
\value{
<<<<<<< HEAD
a simple featurs (sf) object
}
\description{
Joins the retrospective averages to an input NHD object and computes the seasoal and monthly averages.
}
\author{
Mike Johnson
=======
a simple features (sf) object
}
\description{
Joins the retrospective averages to an input NHD object and computes the seasoal, annual and monthly averages.
}
\author{
Mike Johnson and Pat Johnson
>>>>>>> cabb7542fda948e253be880263e733cc398e4e85
}
