% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join nwmRetro data to NHD object}
\usage{
join(obj = NULL)
}
\arguments{
\item{obj}{a NHD object with a COMID attribute}
}
\value{
a simple features (sf) object
}
\description{
Joins the retrospective averages to an input NHD object and computes the seasoal, annual and monthly averages.
}
\author{
Mike Johnson and Pat Johnson
}
