% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwmRetro_data.R
\docType{data}
\name{nwm_averages}
\alias{nwm_averages}
\title{Avergage Flow Conditions}
\format{a data.frame with monthly average stream flow values and timezones}
\usage{
nwm_averages
}
\description{
Dataset averaged Data from the NWM v1.2
}
\examples{
\dontrun{
nwmRetro::nwm_averages
}
}
\keyword{datasets}
